Refractered the Backend Code
