+++ Server/.gitignore
@@ -28,3 +28,8 @@ venv/
 # typescript
 *.tsbuildinfo
 next-env.d.ts
+tests/
+
+__pycache__/
+*.pyc
+*.pyo
